chore: mistakeFeed페이지를 기본 index페이지로 변경
